+++ src/webgl/context.rs
@@ -101,7 +101,7 @@ pub enum CullFace {
 
 #[repr(u32)]
 #[derive(Clone, Copy, Debug)]
-pub enum DrawType {
+pub enum Primitive {
   Lines = WebGl2RenderingContext::LINES,
   Triangles = WebGl2RenderingContext::TRIANGLES,
   TriangleFan = WebGl2RenderingContext::TRIANGLE_FAN,
@@ -334,27 +334,27 @@ impl Context {
   }
 
   #[inline]
-  pub fn draw_arrays(&self, ty: DrawType, count: i32) {
-    let () = self.0.draw_arrays(ty as _, 0, count);
+  pub fn draw_arrays(&self, primitive: Primitive, count: i32) {
+    let () = self.0.draw_arrays(primitive as _, 0, count);
     debug_assert_eq!(self.error(), Ok(()));
   }
 
   #[inline]
-  pub fn draw_arrays_instanced(&self, ty: DrawType, count: i32, instance_count: i32) {
+  pub fn draw_arrays_instanced(&self, primitive: Primitive, count: i32, instance_count: i32) {
     let () = self
       .0
-      .draw_arrays_instanced(ty as _, 0, count, instance_count);
+      .draw_arrays_instanced(primitive as _, 0, count, instance_count);
     debug_assert_eq!(self.error(), Ok(()));
   }
 
   #[inline]
-  pub fn draw_elements<T>(&self, ty: DrawType, count: i32)
+  pub fn draw_elements<T>(&self, primitive: Primitive, count: i32)
   where
     T: BuiltinType,
   {
     let () = self
       .0
-      .draw_elements_with_i32(ty as _, count, T::as_type() as _, 0);
+      .draw_elements_with_i32(primitive as _, count, T::as_type() as _, 0);
     debug_assert_eq!(self.error(), Ok(()));
   }